requirements/edan.txt
Flask-Bootstrap==3.0.3.1
Flask-Moment==0.2.0
SPARQLWrapper==1.5.2
gspread==0.1.0
nltk==2.0.4
numpy==1.8.0
quepy==0.2
rdflib==4.0.1